bump version, drop debian directory as it's unmaintained here. some more thread removal work for stats, video preview merge (not built by default) and a couple of bug fixes from reported feedback.
svn path=/icecast/branches/kh/icecast/; revision=15372

+++ src/refbuf.c
@@ -35,7 +35,7 @@ void refbuf_shutdown(void)
 {
 }
 
-refbuf_t *refbuf_new(size_t size)
+refbuf_t *refbuf_new (unsigned int size)
 {
     refbuf_t *refbuf;
